Add time-series histogram post
Also, update the index and rss templates so drafts are rendered but not
indexed. It looks like the default rss template only includes a post
summary but I was getting the entire post contents in the feed; maybe
the default behaviour changed recently. In any case I changed rss.xml to
put .Content as the item <description> instead of .Summary.